Search improvement
One of the places we pull the sex from (for fashion) is the department
lists that appear in the LHS navigation on most sites. We were
previously searching until we found the first sex and going with that,
which didn’t make sense because if you search just the product name you
usually get both sexes. Now code loops through all departments and, if
it finds both men and women mentioned, returns results for both.