Display only a subset of websites on the options screen
With website level recs coming in, we are going to see a LOT more (very
specific) websites (e.g. British Gas) appearing. It’s silly to display
them all, so let’s filter to just the generic shopping ones. This
needed to be done quickly, so a hardcoded list of IDs have been used.
Sorry.