Upgrade to Foundation 6
This happened kind of by accident, when fixing qtips. The Foundation
upgrade caused all sorts of problems. Joyride had to be removed (not
supported in F6), and the options page required considerable rework to
get the new tabs working. The sidebar grid system also upgraded to F6’s
grid-x format.

The options page especially was done in a rush and has been implemented
poorly. Don’t look at the CSS!

Included in this change, originally in an attempt to fix the qtip
issue, is a generic wait system on all scripts being injected for the
sidebar. So iFrame no longer informs CS_main when it’s loaded. Instead,
CS_main waits for jQuery/foundation/if_main to load before proceeding.
This is probably overkill, but it works.